Made some formatting changes
Things to do before we hand it in:
1) Maybe add a picture of the finished recipe at the top left corner of
the ingredient screen
2) Add the url to access the actual recipe
3) Whatever top-down design stuff we need to do. If you want to improve
it aesthetically or functionality wise go ahead. I'll prob be taking a
break from this for most of tomorrow so the reigns are all yours. Atm
our limitations is that our program can only display 10 ingredients
because there is no scrollbar.