Image display half works
I changed the way its finding the source. Now it's directly a string
instead of being appended from a list. However, the image still sorta
freezes at the loading screen and only finishes loading when u exit the
screen and return back to it which is pretty weird O_O. I think we need
to make add a clock function to repeatedly call or something. Also added
a url open thing that doesn't work